# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.TableModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      number_of_guests:
        type: integer
      table_id:
        type: string
      table_number:
        type: integer
      updated_at:
        type: string
    required:
    - number_of_guests
    - table_number
    type: object
  model.UserModel:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      id:
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      refresh_token:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
host: localhost:3000
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive Restaurant Management System API for managing users,
    menus, food items, orders, tables, invoices, and notes. Available on both local
    and online servers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
  x-servers:
  - description: Local Development Server
    url: http://localhost:3000/api/v1
  - description: Online Production Server
    url: https://restaurant-management-f9kx.onrender.com/api/v1
paths:
...
  /table:
    post:
      consumes:
      - application/json
      description: Create a new restaurant table with number of guests and table number
      parameters:
      - description: Table data (number_of_guests and table_number required)
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/model.TableModel'
      produces:
      - application/json
      responses:
        "200":
          description: Table created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new table
      tags:
      - Table
  /table/{id}:
    delete:
      description: Delete a table by table ID
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Table deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a table
      tags:
      - Table
    get:
      description: Retrieve table details by table ID
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Table details
          schema:
            $ref: '#/definitions/model.TableModel'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific table
      tags:
      - Table
    put:
      consumes:
      - application/json
      description: Update table details by table ID
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated table data
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/model.TableModel'
      produces:
      - application/json
      responses:
        "200":
          description: Table updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a table
      tags:
      - Table
  /tables:
    get:
      description: Retrieve a list of all restaurant tables
      produces:
      - application/json
      responses:
        "200":
          description: List of tables
          schema:
            items:
              $ref: '#/definitions/model.TableModel'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all tables
      tags:
      - Table
  /user:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, phone and personal
        information
      parameters:
      - description: User data (first_name, last_name, email, password, phone required)
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserModel'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully with ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration/sign up
      tags:
      - User
  /user/{id}:
    get:
      description: Retrieve user details by providing the user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/model.UserModel'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific user by ID
      tags:
      - User
  /users:
    get:
      description: Retrieve a list of all users with pagination support
      parameters:
      - description: 'Number of records per page (default: 10)'
        in: query
        name: recordPerPage
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users with pagination
      tags:
      - User
schemes:
- http
- https
swagger: "2.0"