# internal/controller/tableController.go
package controller

import (
	"RESTAURANT-MANAGEMENT/internal/database"
	"RESTAURANT-MANAGEMENT/internal/model"
	"context"
	"log"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/go-playground/validator/v10"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
	"go.mongodb.org/mongo-driver/mongo/options"
)

// CreateTable godoc
// @Summary Create a new table
// @Description Create a new restaurant table with number of guests and table number
// @Tags Table
// @Accept json
// @Produce json
// @Param table body model.TableModel true "Table data (number_of_guests and table_number required)"
// @Success 200 {object} map[string]interface{} "Table created successfully"
// @Failure 400 {object} map[string]string "Bad request - validation error"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /table [post]
func CreateTable() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var tableCollection = database.Collection("tables")

		var table model.TableModel

		if err := c.BindJSON(&table); err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
			return
		}

		validate := validator.New()
		validationErr := validate.Struct(table)

		if validationErr != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": validationErr.Error()})
			return
		}

		table.Created_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
		table.Updated_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))

		table.ID = primitive.NewObjectID()
		table.Table_id = table.ID.Hex()
		result, insertErr := tableCollection.InsertOne(ctx, table)

		if insertErr != nil {
			msg := "Table item was not created"
			c.JSON(http.StatusInternalServerError, gin.H{"error": msg})
			return
		}
		defer cancel()

		c.JSON(http.StatusOK, result)

	}
}

// GetTable godoc
// @Summary Get a specific table
// @Description Retrieve table details by table ID
// @Tags Table
// @Produce json
// @Param id path string true "Table ID"
// @Success 200 {object} model.TableModel "Table details"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /table/{id} [get]
func GetTable() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var tableCollection = database.Collection("tables")
		tableId := c.Param("table_id")
		var table model.TableModel

		err := tableCollection.FindOne(ctx, bson.M{"table_id": tableId}).Decode(&table)
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occure while fetching the tables"})
		}

		c.JSON(http.StatusOK, table)
	}
}

// GetTables godoc
// @Summary Get all tables
// @Description Retrieve a list of all restaurant tables
// @Tags Table
// @Produce json
// @Success 200 {object} []model.TableModel "List of tables"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /tables [get]
func GetTables() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		var tableCollection = database.Collection("tables")

		result, err := tableCollection.Find(context.TODO(), bson.M{})
		defer cancel()

		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occur"})
		}

		var allTables []bson.M
		if err = result.All(ctx, &allTables); err != nil {
			log.Fatal(err)
		}
		c.JSON(http.StatusOK, allTables)
	}
}

// UpdateTable godoc
// @Summary Update a table
// @Description Update table details by table ID
// @Tags Table
// @Accept json
// @Produce json
// @Param id path string true "Table ID"
// @Param table body model.TableModel true "Updated table data"
// @Success 200 {object} map[string]interface{} "Table updated successfully"
// @Failure 400 {object} map[string]string "Bad request"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /table/{id} [put]
func UpdateTable() gin.HandlerFunc {
	return func(c *gin.Context) {
		ctx, cancel := context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var tableCollection = database.Collection("tables")

		var table model.TableModel

		tableId := c.Param("table_id")

		if err := c.BindJSON(&table); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		}

		var updatedObj primitive.D

		if table.Number_of_guests != nil {
			updatedObj = append(updatedObj, bson.E{Key: "number_of_guests", Value: table.Number_of_guests})
		}

		if table.Table_number != nil {
			updatedObj = append(updatedObj, bson.E{Key: "table_number", Value: table.Table_number})
		}

		table.Updated_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))

		opt := options.Update().SetUpsert(true)

		filter := bson.M{"table_id": tableId}

		result, err := tableCollection.UpdateOne(
			ctx,
			filter,
			bson.D{{
				Key:   "$set",
				Value: updatedObj,
			}},
			opt,
		)

		if err != nil {
			msg := "table item update failed"
			c.JSON(http.StatusInternalServerError, gin.H{"error": msg})
			return
		}

		defer cancel()
		c.JSON(http.StatusOK, result)

	}
}

// DeleteTable godoc
// @Summary Delete a table
// @Description Delete a table by table ID
// @Tags Table
// @Produce json
// @Param id path string true "Table ID"
// @Success 200 {object} map[string]string "Table deleted successfully"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /table/{id} [delete]
func DeleteTable() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.JSON(200, gin.H{
			"result": "Delete Table",
		})
	}
}