# internal/controller/userController.go
package controller

import (
	"RESTAURANT-MANAGEMENT/internal/database"
	"RESTAURANT-MANAGEMENT/internal/helper"
	"RESTAURANT-MANAGEMENT/internal/model"
	"context"
	"log"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
	"go.mongodb.org/mongo-driver/mongo"
	"golang.org/x/crypto/bcrypt"
)

// GetUsers godoc
// @Summary Get all users with pagination
// @Description Retrieve a list of all users with pagination support
// @Tags User
// @Produce json
// @Param recordPerPage query int false "Number of records per page (default: 10)"
// @Param page query int false "Page number (default: 1)"
// @Success 200 {object} map[string]interface{} "List of users"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users [get]
func GetUsers() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var userCollection = database.Collection("users")

		recordPerPage, err := strconv.Atoi(c.Query("recordPerPage"))
		if err != nil || recordPerPage < 1 {
			recordPerPage = 10
		}

		page, err1 := strconv.Atoi(c.Query("page"))
		if err1 != nil || page < 1 {
			page = 1
		}

		startIndex := (page - 1) * recordPerPage

		matchStage := bson.D{{Key: "$match", Value: bson.D{}}}
		projectStage := bson.D{
			{Key: "$project", Value: bson.D{
				{Key: "id", Value: 0},
				{Key: "total_count", Value: 1},
				{Key: "user_items", Value: bson.D{{Key: "$slice", Value: []interface{}{"$date", startIndex, recordPerPage}}}},
			}},
		}

		result, err := userCollection.Aggregate(ctx, mongo.Pipeline{
			matchStage,
			projectStage,
		})

		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occured while listing to user item"})
			return
		}

		var allUsers []bson.M
		if err = result.All(ctx, &allUsers); err != nil {
			log.Fatal(err)
		}

		c.JSON(http.StatusOK, allUsers[0])

		//either pass an error

		//identify want to return all users based on the various query parsed
	}
}

// GetUsers godoc
// @Summary Get all users with pagination
// @Description Retrieve a list of all users with pagination support
// @Tags User
// @Produce json
// @Param recordPerPage query int false "Number of records per page (default: 10)"
// @Param page query int false "Page number (default: 1)"
// @Success 200 {object} map[string]interface{} "List of users"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users [get]
func GetUser() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var userCollection = database.Collection("users")

		userId := c.Param("user_id")
		var user model.UserModel

		err := userCollection.FindOne(ctx, bson.M{"user_id": userId}).Decode(&user)

		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occured while listing to user items"})
		}

		c.JSON(http.StatusOK, user)
	}
}

// GetUser godoc
// @Summary Get a specific user by ID
// @Description Retrieve user details by providing the user ID
// @Tags User
// @Produce json
// @Param id path string true "User ID"
// @Success 200 {object} model.UserModel "User details"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /user/{id} [get]
func UpdateUser() gin.HandlerFunc {
	return func(c *gin.Context) {
		// var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)

	}
}

// UpdateUser godoc
// @Summary Update user information
// @Description Update specific user details by user ID
// @Tags User
// @Accept json
// @Produce json
// @Param id path string true "User ID"
// @Param user body model.UserModel true "Updated user data"
// @Success 200 {object} map[string]interface{} "User updated successfully"
// @Failure 400 {object} map[string]string "Bad request"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /user-update/{id} [put]

// SignUp godoc
// @Summary User registration/sign up
// @Description Create a new user account with email, password, phone and personal information
// @Tags User
// @Accept json
// @Produce json
// @Param user body model.UserModel true "User data (first_name, last_name, email, password, phone required)"
// @Success 200 {object} map[string]interface{} "User created successfully with ID"
// @Failure 400 {object} map[string]string "Bad request - validation error"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /user [post]
func SignUp() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var userCollection = database.Collection("users")
		var user model.UserModel

		//convert the JSON data coming from postman to something that golang understand
		if err := c.BindJSON(&user); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		}
		//validate the data based on user struct
		validationErr := validate.Struct(user)
		if validationErr != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": validationErr.Error()})
			return
		}

		//you'll check if the email has already been used by another user
		count, err := userCollection.CountDocuments(ctx, bson.M{"email": user.Email})
		defer cancel()

		if err != nil {
			log.Panic(err)
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occured while checking for the email"})
			return
		}

		//hash password
		password := HashPasssword(*user.Password)
		user.Password = &password

		//you'll also check if the phone no. has already been used by another
		count, err = userCollection.CountDocuments(ctx, bson.M{"phone": user.Phone})

		defer cancel()

		if err != nil {
			log.Panic(err)
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occured while checking for the phone"})
			return
		}

		if count > 0 {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "This email or phone number already exists"})
		}
		//create some extra details for the user object - created_at, updated_at, ID
		user.Created_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
		user.Updated_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))

		user.ID = primitive.NewObjectID()
		hexID := user.ID.Hex()
		user.User_id = &hexID

		//generate token and refresh token (generate all token function from helper )
		token, refreshToken, err := helper.GenerateAllToken(
			*user.Email,
			*user.First_name,
			*user.Last_name,
			*user.User_id,
		)

		if err != nil {
			log.Panic(err)
			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occured while generating token"})
			return
		}

		user.Token = &token
		user.Refresh_Token = &refreshToken
		//if all ok, then you insert this new user into the user collection
		resultInsertionNumber, insertErr := userCollection.InsertOne(ctx, user)
		if insertErr != nil {
			msg := "user item was not created"
			c.JSON(http.StatusInternalServerError, gin.H{"error": msg})
			return
		}
		defer cancel()

		//return status ok and send the result back
		c.JSON(http.StatusOK, resultInsertionNumber)
	}
}

func LogIn() gin.HandlerFunc {
	return func(c *gin.Context) {
		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
		defer cancel()
		var userCollection = database.Collection("users")
		var user model.UserModel
		var foundUser model.UserModel

		//convert the login data from postman which is in JSON to golang readable format
		if err := c.BindJSON(&user); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		}

		//find a user with that email and see if that user exists
		err := userCollection.FindOne(ctx, bson.M{"email": user.Email}).Decode(&user)
		defer cancel()
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "user not found, login seems to be incorrect"})
			return
		}

		//then you will verify the password
		passwordIsValid, msg := VerifyPassword(*user.Password, *foundUser.Password)
		if passwordIsValid != true {
			c.JSON(http.StatusInternalServerError, gin.H{"error": msg})
			return
		}
		//if all goes well, then you'll generate tokens
		token, refreshToken, _ := helper.GenerateAllToken(*foundUser.Email, *foundUser.First_name, *foundUser.Last_name, *foundUser.User_id)

		//update tokens - token and refresh token
		helper.UpdateAllTokens(token, refreshToken, *foundUser.User_id)

		//return statusOk
		c.JSON(http.StatusOK, foundUser)
	}
}

func HashPasssword(password string) string {
	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
	if err != nil {
		log.Panic(err)
	}

	return string(bytes)
}

func VerifyPassword(userPassword string, providePassword string) (bool, string) {
	err := bcrypt.CompareHashAndPassword([]byte(providePassword), []byte(userPassword))
	check := true
	msg := ""
	if err != nil {
		msg = "login or password is incorrect"
		check = false
	}
	return check, msg
}